feat(data): a decimal type, because storing quantities as floats undid the rounding
Odoo splits storage from arithmetic: a quantity or a price is stored as exact
decimal and computed as a binary float, with float_round and float_compare
standing between the two. The first version here copied the arithmetic and missed
the storage, so quantities went into DOUBLE PRECISION — where 0.1 comes back as
0.1000000000000000055, and every trip through the database puts back the error
the rounding had just taken out.

decimal is now its own field type:

- Postgres: NUMERIC, unbounded, as Odoo uses for quantities and money
- SQLite: TEXT, because SQLite has no exact decimal at all — NUMERIC affinity
  silently becomes REAL, so text is the only storage that returns what it was
  given
- Both adapters hand it back as a string, and ctx converts, since it is the one
  place that knows both the model and the row. Arithmetic stays on numbers.

uom.Unit stores factor and rounding as decimal. A factor of 0.001 in a double is
0.001000000000000000020816681711721685, and every conversion through it carries
that into a stock figure.

Tested both ways: awkward values round-trip unchanged through SQLite and through
live Postgres, the column type is asserted to be numeric, and the raw value is
confirmed to be "0.1" rather than a binary approximation of it.

227/227 tests, 9/9 against live Postgres, 0 type errors, 9/9 type assertions.

+++ packages/ketsuite/src/modules/uom/models.ts
@@ -30,8 +30,11 @@ export const models: Record<string, ModelDef> = {
       categoryId: 'ref:uom.Category',
       // 'reference' | 'bigger' | 'smaller' — validated on write, see functions.ts
       type: 'text',
-      factor: 'float',
-      rounding: 'float',
+      // decimal, not float: a factor of 0.001 written to a double comes back as
+      // 0.001000000000000000020816681711721685, and every conversion through it
+      // carries that error into a stock figure.
+      factor: 'decimal',
+      rounding: 'decimal',
       active: 'bool',
     },
   },
